#23 extract Node class. Add null node test
