+++ src/main/java/com/raychenon/hackerrank/tree/Node.java
@@ -1,0 +1,32 @@
+package com.raychenon.hackerrank.tree;
+
+/**
+ * User: raychenon
+ * Date: 6/2/19
+ * https://www.hackerrank.com/challenges/tree-postorder-traversal/problem
+ */
+public class Node {
+    int data;
+    Node left;
+    Node right;
+
+    public Node(int data){
+        this.data = data;
+    }
+
+    public Node insert(Node root, int data){
+        if(root == null) {
+            return new Node(data);
+        } else {
+            Node cur;
+            if(data <= root.data) {
+                cur = insert(root.left, data);
+                root.left = cur;
+            } else {
+                cur = insert(root.right, data);
+                root.right = cur;
+            }
+            return root;
+        }
+    }
+}
